Resolved conflicts: kept existing CollisionSystem API (audioService param,
linear ball search) and multi-ABI native copy task. Took ProGuard enablement,
proguard-rules.pro, and signing config template from the agent.
Ball entity caching (#36) deferred — linear search is sufficient for now.

+++ core/src/main/kotlin/com/streetsoccer/ecs/systems/CollisionSystem.kt
@@ -214,8 +214,8 @@ class CollisionSystem(
 
                 if (transform.height <= STOPPED_HEIGHT_THRESHOLD) {
                     val speed = kotlin.math.sqrt(
-                        velocity.vx * velocity.vx + velocity.vy * velocity.vy + velocity.vz * velocity.vz
-                    )
+                        (velocity.vx * velocity.vx + velocity.vy * velocity.vy + velocity.vz * velocity.vz).toDouble()
+                    ).toFloat()
                     if (speed < STOPPED_SPEED_THRESHOLD) {
                         sessionAccumulator.breakStreak()
                         gameStateManager.transitionTo(GameState.ImpactMissed)
